logos-sandbox/pyproject.toml
[project]
name = "logos-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "fastapi[standard]>=0.115.12",
    "instructor>=1.8.1",
    "mcp>=1.8.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.1",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.2",
]
